👷 use a dedicated dev server script
